Move RandomizedRollingUpgradeIT into logsdb/qa/rolling-upgrade
